update to latest svelte kit version
