+++ src/app.html
@@ -1,18 +1,18 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en">
 	<head>
 		<meta charset="utf-8" />
 		<meta name="description" content="Twitch Text to Speech" />
-		<link rel="icon" href="%svelte.assets%/favicon.png" />
+		<link rel="icon" href="%sveltekit.assets%/favicon.png" />
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
 
 		<link
 			href="https://fonts.googleapis.com/css?family=Hind|Open+Sans&display=swap"
 			rel="stylesheet"
 		/>
-		%svelte.head%
+		%sveltekit.head%
 	</head>
-	<body>
-		<div>%svelte.body%</div>
+	<body data-sveltekit-preload-data="hover">
+		<div style="display: contents">%sveltekit.body%</div>
 	</body>
 </html>
