+++ src/routes/+layout.svelte
@@ -1,5 +1,5 @@
 <script>
-	import '../global.css';
+	import '../app.css';
 </script>
 
 <slot />
